Update address.yaml
So looking at the xsd schema they define the internal types with adb references and they use the namespace http://namespaces.geoplace.co.uk/addressbase/2.1 to map these.  Therefore I think we should follow this for now.  Need to investigate uprn as it is listed as an integer in the specification, but we may be passing around as a string.