SPT-624 Updated schema to represent existing CRI data
In order for us to enable the full schema checking within SPOT we need to update
the data-vocab to reflect the current data being sent to SPOT from CRIs. If we
enavble full schema checking within SPOT with the current schema a large majority
of requests would fail against the current schema.