BAU: Remove `v` prefix when publishing packages
When publishing packages, ensure the `v` prefix is removed from the
version name (the `v` should be retained while creating the GitHub
release).

The inclusion of the `v` is proving troublesome with the Maven
repository, NPM is clever enough to remove `v` but the Gradle publish
action just uses the version verbatim and downstream this causes
Dependabot to do strange things.