SPT-1296: Added code to remove duplicate fields
Added code to review duplicate fields. This will remove fields which are
both matching the parents definition and also fields which inherit from
the parents definition. If the redefined field does not inherit from the
parent then it is left, it will then be up to the end user to determine
what they want to do with the duplicate field.

Later PR will add generics to base classes, so that they return the
correct type.