Remove LinkML as a cached dependency and use live
This is a proposal, I hope to have a good debate on the PR.
We made a decision a while back to not use the main release branch of
LinkML, as a result we forked it at a point in time and then loaded
LinkML manually into this repo.

Then we stopped maintaining our fork.

The Fork [1] is 2070 commits behind main, has not been udpated in 3
years, and is full of vulnerabilities.

The vulns are not actually likely to be a source of risk given how we
use this to build a static site, but the risk is not zero.

I propose we revert this model, go back to usual supply chain approaches
and archive the fork.

Anyone who can dig out the reasons we did this way back when would be
great to review those before we merge

[1]: https://github.com/govuk-one-login/linkml